refactor: board 관련 repository, service, controller 수정 (#70)
