+++ src/main/java/org/example/stockradar/feature/board/dto/BoardResponseDto.java
@@ -1,0 +1,13 @@
+package org.example.stockradar.feature.board.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class BoardResponseDto {
+}
