+++ src/main/java/org/example/stockradar/feature/board/entity/BoardContent.java
@@ -5,9 +5,7 @@ import lombok.*;
 
 @Entity
 @Getter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class BoardContent {
 
     @Id
