+++ src/main/java/org/example/stockradar/feature/board/entity/Comments.java
@@ -10,9 +10,7 @@ import java.util.List;
 
 @Entity
 @Getter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class Comments {
 
     @Id
