+++ src/main/java/org/example/stockradar/feature/test/testController.java
@@ -83,10 +83,10 @@ public class testController {
     public String productCategory() {return "/product/productCategory";}
 
     //커뮤니티 이동
-    @GetMapping("/board/boardList")
-    public String boardMain() {
-        return "board/boardList";
-    }
+//    @GetMapping("/board/boardList")
+//    public String boardMain() {
+//        return "board/boardList";
+//    }
 
     @GetMapping("/product/gpu/rtx3050")
     public String gpuRtx3050() {return "/product/gpu/rtx3050";}
